MessagingQueue/producers/producer1/requirements.txt
asyncio
stomp
aiohttp
ssl